INGM-708 move no_error_tracker
